% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{annotateTSS}
\alias{annotateTSS}
\title{Annotate the provided Transcription Start Sites}
\usage{
annotateTSS(species, genome)
}
\arguments{
\item{species}{String indicating the name of the species.It can be either "Homo sapiens","Mus musculus".}

\item{genome}{String indicating the name of the genome assembly.It can be either "hg38",
"hg19", "mm10", "mm9".}
}
\value{
A data.frame with each gene TSS locus
}
\description{
This function annotates transcription start sites based on species and genome versions
}
\examples{
# TSS annotations for the hg38
tssdata <- annotateTSS("Homo sapiens", "hg38")
head(tssdata)

# TSS annotations for the mm10
tssdata <- annotateTSS("Mus musculus", "mm10")
head(tssdata)
}
