% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{calculateHiCWeights}
\alias{calculateHiCWeights}
\title{Calculate promoters, enhancers, and Hi-C matrix for genes}
\usage{
calculateHiCWeights(
  SCEGdata,
  species,
  genome,
  focus_gene,
  averHicPath,
  TSSwindow = 1000,
  upstream = 250000,
  downstream = 250000,
  verbose = TRUE
)
}
\arguments{
\item{SCEGdata}{Preprocessed data for SCEG-HiC.}

\item{species}{Character string specifying the species name. Supported values are "Homo sapiens" or "Mus musculus".}

\item{genome}{Character string specifying the genome assembly. Supported values are "hg38", "hg19", "mm10", or "mm9".}

\item{focus_gene}{Character vector of gene names to focus on.}

\item{averHicPath}{Path to the bulk average Hi-C data.}

\item{TSSwindow}{Numeric specifying the number of base pairs to extend upstream and downstream around each TSS to define promoters. Default is 1000 bp (total window size of 2000 bp).}

\item{upstream}{Numeric specifying the number of base pairs upstream of each TSS to define enhancers. Default is 250,000 bp (250 kb).}

\item{downstream}{Numeric specifying the number of base pairs downstream of each TSS to define enhancers. Default is 250,000 bp (250 kb).}

\item{verbose}{Logical. Should progress messages and warnings be printed?}
}
\value{
A named list where each element corresponds to a focus gene and contains:
\itemize{
\item \code{promoters}: data.frame of promoter regions.
\item \code{enhancers}: data.frame of enhancer regions.
\item \code{contact}: bulk average Hi-C contact matrix between the gene and enhancers.
}
}
\description{
This function obtains enhancers and promoters of each gene, and calculates the bulk average Hi-C matrix of contacts between the gene and its enhancers.
}
\examples{
data(multiomic_small)
SCEGdata <- process_data(multiomic_small, k_neigh = 5, max_overlap = 0.5)
fpath <- system.file("extdata", package = "SCEGHiC")
gene <- c("TRABD2A", "GNLY", "MFSD6", "CTLA4", "LCLAT1", "NCK2", "GALM", "TMSB10", "ID2", "CXCR4")
weight <- calculateHiCWeights(SCEGdata, species = "Homo sapiens", genome = "hg38", focus_gene = gene, averHicPath = fpath)
weight[["CTLA4"]]$promoters
weight[["CTLA4"]]$enhancers
head(weight[["CTLA4"]]$contact)

}
