% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{connections_Plot}
\alias{connections_Plot}
\title{Plot links of SCEG-HiC on focused genes}
\usage{
connections_Plot(
  SCEG_HiC_Result,
  species,
  genome,
  focus_gene,
  cutoff = NULL,
  upstream = 250000,
  downstream = 250000,
  gene_anno = NULL
)
}
\arguments{
\item{SCEG_HiC_Result}{SCEG_HiC result}

\item{species}{String indicating the name of the species.It can be either "Homo sapiens","Mus musculus".}

\item{genome}{String indicating the name of the genome assembly.It can be either "hg38",
"hg19", "mm10", "mm9".}

\item{focus_gene}{The focused a gene.}

\item{cutoff}{The threshold of gene-peak.Default is NULL.If \code{"aggregate"} is TRUE ,we recommend the cutoff of 0.01.
If \code{"aggregate"} is FALSE ,we recommend the cutoff of 0.001.}

\item{upstream}{Numeric specifying the window size (in base pairs) to pad around upstream of each TSS, to fetch gene's enhancers.
Default is 250 kb (so 500 kb window is drawn arund each TSS).}

\item{downstream}{Numeric specifying the window size (in base pairs) to pad around downstream of each TSS, to fetch gene's enhancers.
Default is 250 kb (so 500 kb window is drawn arund each TSS).}

\item{gene_anno}{Either \code{NULL} or a data.frame. The data.frame should be in a form compatible with the Gviz function.
\code{\link[Gviz]{GeneRegionTrack-class}} (cannot have NA as column names).}
}
\value{
A gene region and gene-peak pairs plot
}
\description{
Plot links of SCEG-HiC on focused genes
}
\examples{
#' data(multiomic_small)
SCEGdata <- process_data(multiomic_small, k_neigh = 5, max_overlap = 0.5)
fpath <- system.file("extdata", package = "SCEGHiC")
gene <- c("TRABD2A", "GNLY", "MFSD6", "CTLA4", "LCLAT1", "NCK2", "GALM", "TMSB10", "ID2", "CXCR4")
weight <- calculateHiCWeights(SCEGdata, species = "Homo sapiens", genome = "hg38", focus_gene = gene, averHicPath = fpath)
results_SCEGHiC <- Run_SCEG_HiC(SCEGdata, weight, focus_gene = gene)
connections_Plot(results_SCEGHiC, species = "Homo sapiens", genome = "hg38", focus_gene = "CTLA4", cutoff = 0.01, gene_anno = NULL)
}
