% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{generate_aggregated_datasets}
\alias{generate_aggregated_datasets}
\title{Create aggregated data for a certain cluster}
\usage{
generate_aggregated_datasets(
  object,
  cell_coord,
  rna_assay = "RNA",
  atac_assay = "peaks",
  k_neigh = 50,
  atacbinary = TRUE,
  max_overlap = 0.8,
  seed = 123,
  verbose = TRUE
)
}
\arguments{
\item{object}{A Seurat object.}

\item{cell_coord}{A similarity matrix or dimensionality reduction (e.g., PCA, UMAP) used for identifying neighbors.}

\item{rna_assay}{Character. Name of the assay containing gene expression data.}

\item{atac_assay}{Character. Name of the assay containing peak (chromatin accessibility) data.}

\item{k_neigh}{Integer. Number of neighboring cells to aggregate per group (default is 50).}

\item{atacbinary}{Logical. Should the aggregated scATAC-seq matrix be binarized?}

\item{max_overlap}{Numeric. Maximum allowed overlap ratio between two aggregated groups (default is 0.8).}

\item{seed}{Integer. Random seed.}

\item{verbose}{Logical. Logical. Should progress messages and warnings be printed?}
}
\value{
A matrix or sparse matrix containing aggregated accessibility or expression values.
}
\description{
Function to generate aggregated inputs of a cetrain cluster. \code{generate_aggregated_datasets}
takes as input sparse data. This function will aggregate binary accessibility scores (or gene expression)
per cell cluster, if they do not overlap any existing group with more than 50\% cells.
}
