% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{generate_aggregated_datastes}
\alias{generate_aggregated_datastes}
\title{Create aggregated data for a certain cluster}
\usage{
generate_aggregated_datastes(
  object,
  cell_coord,
  rna_assay = "RNA",
  atac_assay = "peaks",
  k_neigh = 50,
  atacbinary = TRUE,
  max_overlap = 0.8,
  seed = 123,
  verbose = TRUE
)
}
\arguments{
\item{object}{object Seurat object.}

\item{cell_coord}{cell_coord similarity matrix or dimiension reductions.}

\item{rna_assay}{Name of assay containing gene expression information.}

\item{atac_assay}{Name of assay containing peak information.}

\item{k_neigh}{Number of cells to aggregate per group.}

\item{atacbinary}{Logical, whether the aggregated scATAC-seq data need binary.}

\item{max_overlap}{The maximum overlapping ratio of two groups.}

\item{seed}{Random seed.}

\item{verbose}{verbose Logical, should warning and info messages be printed?}
}
\value{
Aggregated data.
}
\description{
Function to generate aggregated inputs of a cetrain cluster. \code{generate_aggregated_datastes}
takes as input sparse data. This function will aggregate binary accessibility scores (or gene expression)
per cell cluster, if they do not overlap any existing group with more than 50\% cells.
}
