% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19TSSdata}
\alias{hg19TSSdata}
\title{hg19TSSdata}
\format{
A data frame with 43327 rows and 3 variables:
\describe{
\item{chr}{Chromosome name (e.g., "chr1").}
\item{TargetGene}{Gene name associated with the TSS.}
\item{TargetGeneTSS}{Position of the TSS.}
}
}
\source{
UCSC RefGene database (\url{https://hgdownload2.soe.ucsc.edu/goldenPath/hg19/bigZips/genes/hg19.refGene.gtf.gz}).
}
\usage{
hg19TSSdata
}
\description{
Transcription start site (TSS) data for the human genome (hg19).
}
\details{
This dataset contains transcription start site annotations for the hg19 human genome assembly.
}
\examples{
data(hg19TSSdata)
head(hg19TSSdata)
}
\keyword{datasets}
