% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38TSSdata}
\alias{hg38TSSdata}
\title{hg38TSSdata}
\format{
A data frame with 47982 rows and 3 variables:
\describe{
\item{chr}{Chromosome name (e.g., "chr1").}
\item{TargetGene}{Gene name associated with the TSS.}
\item{TargetGeneTSS}{Position of the TSS.}
}
}
\source{
UCSC RefGene database (\url{https://hgdownload2.soe.ucsc.edu/goldenPath/hg38/bigZips/genes/hg38.refGene.gtf.gz}).
}
\usage{
hg38TSSdata
}
\description{
Transcription start site (TSS) data for the human genome (hg38).
}
\details{
This dataset contains transcription start site annotations for the hg38 human genome assembly.
}
\examples{
data(hg38TSSdata)
head(hg38TSSdata)
}
\keyword{datasets}
