% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{liftOverPeaks}
\alias{liftOverPeaks}
\title{Peaks corresponding genome version conversion}
\usage{
liftOverPeaks(split_peaks, genome)
}
\arguments{
\item{split_peaks}{Peaks of genome coordinate conversion.}

\item{genome}{String indicating the name of the genome assembly.It can be either "hg38","mm9".}
}
\value{
A data.frame with peaks after genome coordinate conversion.
}
\description{
The function convert genomic coordinates based on provided peaks.\code{SCEGHIC} provides the bulk
average Hi-C genome versions of hg19 and mm10, so   require genome coordinate conversion for hg38 and mm10.
}
\examples{
peaks <- c(
  "chr2-973888-974755", "chr2-3195868-3196264", "chr2-3242398-3243508", "chr2-6863856-6866834", "chr2-6913141-6914394",
  "chr2-6924663-6924972", "chr2-6999095-6999554", "chr2-7007713-7008819", "chr2-7083945-7084032", "chr2-7097148-7098029"
)
split_peaks <- do.call("rbind", strsplit(as.character(peaks), "-", fixed = TRUE))
peakinfo <- liftOverPeaks(split_peaks, "hg38")
head(peakinfo)
}
