% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm10TSSdata}
\alias{mm10TSSdata}
\title{mm10TSSdata}
\format{
A data frame with 31599 rows and 3 variables:
\describe{
\item{chr}{Chromosome name (e.g., "chr1").}
\item{TargetGene}{Gene name associated with the TSS.}
\item{TargetGeneTSS}{Position of the TSS.}
}
}
\source{
UCSC RefGene database (\url{https://hgdownload2.soe.ucsc.edu/goldenPath/mm10/bigZips/genes/mm10.refGene.gtf.gz}).
}
\usage{
mm10TSSdata
}
\description{
Transcription start site (TSS) data for the mouse genome (mm10).
}
\details{
This dataset contains transcription start site annotations for the mm10 mouse genome assembly.
}
\examples{
data(mm10TSSdata)
head(mm10TSSdata)
}
\keyword{datasets}
