% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm9TSSdata}
\alias{mm9TSSdata}
\title{mm9TSSdata}
\format{
A data frame with 31479 rows and 3 variables:
\describe{
\item{chr}{Chromosome name (e.g., "chr1").}
\item{TargetGene}{Gene name associated with the TSS.}
\item{TargetGeneTSS}{Position of the TSS.}
}
}
\source{
UCSC RefGene database (\url{https://hgdownload2.soe.ucsc.edu/goldenPath/mm9/bigZips/genes/mm9.refGene.gtf.gz}).
}
\usage{
mm9TSSdata
}
\description{
Transcription start site (TSS) data for the mouse genome (mm9).
}
\details{
This dataset contains transcription start site annotations for the mm9 mouse genome assembly.
}
\examples{
data(mm9TSSdata)
head(mm9TSSdata)
}
\keyword{datasets}
