% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multiomic_small}
\alias{multiomic_small}
\title{A small example paired scRNA-seq and scATAC-seq data.}
\format{
A Seurat object with the following assays
\describe{
\item{RNA}{A gene x cell dataset}
\item{peaks}{A peak x cell dataset}
}
}
\source{
\url{https://support.10xgenomics.com/single-cell-multiome-atac-gex/datasets/1.0.0/pbmc_granulocyte_sorted_10k}
}
\usage{
multiomic_small
}
\description{
A subsetted version of 10x Genomics 10k human (hg38) PBMC paired scRNA-seq and scATAC-seq dataset
}
\concept{data}
\keyword{datasets}
