% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{process_data}
\alias{process_data}
\title{Create preprocessed input data for SCEG-HiC}
\usage{
process_data(
  object,
  aggregate = TRUE,
  celltype = NULL,
  rna_assay = "RNA",
  atac_assay = "peaks",
  cellnames = NULL,
  atacbinary = TRUE,
  k_neigh = 50,
  max_overlap = 0.8,
  size_factor_normalize = TRUE,
  reduction.name = NULL,
  seed = 123,
  verbose = TRUE
)
}
\arguments{
\item{object}{A Seurat object.}

\item{aggregate}{Logical. Whether to generate aggregated (clustered) data.
If \code{FALSE}, a specific \code{celltype} must be provided.}

\item{celltype}{Character. A specific cell type to subset for non-aggregated data.
Required if \code{aggregate = FALSE}.}

\item{rna_assay}{Character. Name of the assay containing gene expression data.
If \code{aggregate = TRUE}, this should contain raw expression counts.
If \code{aggregate = FALSE}, this should contain normalized expression values.}

\item{atac_assay}{Character. Name of the assay containing chromatin accessibility data.
If \code{aggregate = TRUE}, this should contain raw accessibility counts.
If \code{aggregate = FALSE}, this should contain normalized accessibility values.}

\item{cellnames}{Character vector. Name(s) of one or more metadata columns used to group the cells.
Default is the current cell identities.}

\item{atacbinary}{Logical. Should the aggregated scATAC-seq matrix be binarized?}

\item{k_neigh}{Integer. Number of cells to aggregate per group (default is 50).}

\item{max_overlap}{Numeric. Maximum allowed overlap ratio between two aggregated groups (default is 0.8).}

\item{size_factor_normalize}{Logical. Whether to normalize accessibility values using size factors.}

\item{reduction.name}{Character. Name of the dimensionality reduction to use for extracting cell coordinates for aggregation.}

\item{seed}{Integer. Random seed.}

\item{verbose}{Logical. Should progress messages and warnings be printed?}
}
\value{
A list containing preprocessed data for SCEG-HiC:
\itemize{
\item If \code{aggregate = TRUE}, returns a list with aggregated RNA and ATAC matrices.
\item If \code{aggregate = FALSE}, returns RNA and ATAC matrices corresponding to the selected cell type.
}
}
\description{
This function generates preprocessed input data for the SCEG-HiC framework. The output can be
directly used by \code{calculateHiCWeights} and \code{Run_SCEG_HiC}. It supports both aggregated
(clustered) data and non-aggregated (cell-type-specific) data generation.
}
\examples{
data(multiomic_small)

# Aggregated data with paired scRNA-seq and scATAC-seq
SCEGdata <- process_data(multiomic_small, k_neigh = 5, max_overlap = 0.5)

# Single-cell type data for cell type "1"
SCEGdata <- process_data(multiomic_small, aggregate = FALSE, celltype = "1")

}
