% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{Run_SCEG_HiC}
\alias{Run_SCEG_HiC}
\title{Run the SCEG-HiC Model}
\usage{
Run_SCEG_HiC(
  SCEGdata,
  HiCWeights,
  method = "wglasso",
  focus_gene,
  normalizeMethod = "1",
  cutoff = NULL,
  verbose = TRUE,
  alpha = NULL
)
}
\arguments{
\item{SCEGdata}{A list containing data preprocessed by \code{SCEG-HiC}.}

\item{HiCWeights}{A data frame of raw bulk average Hi-C interaction weights.
Output of \code{calculateHiCWeights()}.}

\item{method}{The method used for gene network reconstruction. Default is \code{"wglasso"}.}

\item{focus_gene}{A character vector of gene symbols to focus on.}

\item{normalizeMethod}{Method used to normalize the average Hi-C scores. Default is "1".
\itemize{
\item \code{"1"}: Normalize by rank scores.
\item \code{"2"}: Normalize by -log10 transformation.
\item \code{"3"}: Binarize the scores (values > 0 become 1; values ≤ 0 become 0).
\item \code{"4"}: Normalize by min-max scaling.
}}

\item{cutoff}{Threshold for selecting gene-peak pairs. Default is \code{NULL}.
If \code{aggregate = TRUE}, we recommend setting \code{cutoff = 0.01}.
If \code{aggregate = FALSE}, we recommend \code{cutoff = 0.001}.}

\item{verbose}{Logical. Should progress messages and warnings be printed?}

\item{alpha}{Multiplicative scaling factor for the penalty parameter (alpha * rho).}
}
\value{
A \code{data.frame} containing correlation values and metadata for each gene-peak pair.
}
\description{
This function runs the SCEG-HiC model to infer gene-enhancer (gene-peak) links using preprocessed
single-cell multi-omics data and bulk average Hi-C data as a prior. It calculates a partial correlation
matrix via the weighted graphical lasso (wglasso) approach and selects gene-peak pairs accordingly.
}
\examples{
data(multiomic_small)
SCEGdata <- process_data(multiomic_small, k_neigh = 5, max_overlap = 0.5)
fpath <- system.file("extdata", package = "SCEGHiC")
gene <- c("TRABD2A", "GNLY", "MFSD6", "CTLA4", "LCLAT1", "NCK2", "GALM", "TMSB10", "ID2", "CXCR4")
weight <- calculateHiCWeights(SCEGdata, species = "Homo sapiens", genome = "hg38", focus_gene = gene, averHicPath = fpath)
results_SCEGHiC <- Run_SCEG_HiC(SCEGdata, weight, focus_gene = gene)
results_SCEGHiC[results_SCEGHiC$gene == "CTLA4", ]

}
