% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{wglasso}
\alias{wglasso}
\title{Gene network reconstruction using prior biological knowledge}
\usage{
wglasso(S, sample, normhicMatrix, element, alpha = NULL)
}
\arguments{
\item{S}{A symmetric p-by-p covariance matrix representing co-variation between elements.}

\item{sample}{An integer indicating the number of samples used to estimate the covariance matrix.}

\item{normhicMatrix}{A normalized average Hi-C contact matrix used as prior biological knowledge for penalization.}

\item{element}{A character vector containing the row/column names (element IDs) of the Hi-C matrix.}

\item{alpha}{Multiplicative scaling factor for the penalty parameter (alpha * rho).}
}
\value{
A symmetric matrix of partial correlation coefficients between elements.
}
\description{
This function reconstructs a gene (or regulatory element) network by integrating prior biological knowledge into the estimation of the precision matrix.
Instead of applying a uniform penalty to all element pairs, this method assigns different penalty values based on prior knowledge regarding whether two elements should be connected.
Such prior information can be derived or obtained from biological data, such as a normalized bulk average Hi-C matrix.
}
