% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{Aggregate_datasets}
\alias{Aggregate_datasets}
\title{Create aggregated data for SCEG-HiC}
\usage{
Aggregate_datasets(
  object,
  k_neigh = 50,
  rna_assay = "RNA",
  atac_assay = "peaks",
  cellnames = NULL,
  atacbinary = TRUE,
  max_overlap = 0.8,
  reduction.name = NULL,
  size_factor_normalize = TRUE,
  seed = 123,
  verbose = TRUE
)
}
\arguments{
\item{object}{A Seurat object.}

\item{k_neigh}{Integer. Number of cells to aggregate per group (default is 50).}

\item{rna_assay}{Character. Name of the assay containing gene expression data.}

\item{atac_assay}{Character. Name of the assay containing peak (chromatin accessibility) data.}

\item{cellnames}{Character vector. Name(s) of one or more metadata columns used to group the cells.
Default is the current cell identities.}

\item{atacbinary}{Logical. Should the aggregated scATAC-seq matrix be binarized?}

\item{max_overlap}{Numeric. Maximum allowed overlap ratio between two aggregated groups (default is 0.8).}

\item{reduction.name}{Character. Name of the dimensionality reduction to use for extracting cell coordinates for aggregation.}

\item{size_factor_normalize}{Logical. Whether to normalize accessibility values using size factors.}

\item{seed}{Integer. Random seed.}

\item{verbose}{Logical. Should progress messages and warnings be printed?}
}
\value{
Aggregated Seurat object.
}
\description{
Generate aggregated inputs for the SCEG-HiC model using the
\code{Aggregate_datasets} function.
}
