% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{annotateTSS}
\alias{annotateTSS}
\title{Annotate transcription start sites (TSS)}
\usage{
annotateTSS(species, genome)
}
\arguments{
\item{species}{Character string specifying the species name. Supported values are "Homo sapiens" or "Mus musculus".}

\item{genome}{Character string specifying the genome assembly. Supported values are "hg38", "hg19", "mm10", or "mm9".}
}
\value{
A data.frame containing TSS annotations filtered by species and genome, including chromosome,
gene name, and TSS genomic loci information.
}
\description{
This function annotates transcription start sites based on the specified species and genome assembly version.
}
\examples{
# TSS annotations for the hg38
tssdata <- annotateTSS("Homo sapiens", "hg38")
head(tssdata)

# TSS annotations for the mm10
tssdata <- annotateTSS("Mus musculus", "mm10")
head(tssdata)
}
