% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{calculateHiCWeights}
\alias{calculateHiCWeights}
\title{Calculate  promoters, enhancers, and Hi-C matrix of genes}
\usage{
calculateHiCWeights(
  SCEGdata,
  species,
  genome,
  focus_gene,
  averHicPath,
  TSSwindow = 1000,
  upstream = 250000,
  downstream = 250000,
  verbose = TRUE
)
}
\arguments{
\item{SCEGdata}{Preprocessed data for SCEG-HiC}

\item{species}{String indicating the name of the species.It can be either "Homo sapiens","Mus musculus".}

\item{genome}{String indicating the name of the genome assembly.It can be either "hg38",
"hg19", "mm10", "mm9".}

\item{focus_gene}{The focused genes.}

\item{averHicPath}{Path to the bulk average Hi-C.}

\item{TSSwindow}{Numeric specifying the window size (in base pairs) to pad around either side of each TSS, to fetch gene's promoters.
Default is 1000 bp (so 2000 bp window is drawn arund each TSS).}

\item{upstream}{Numeric specifying the window size (in base pairs) to pad around upstream of each TSS, to fetch gene's enhancers.
Default is 250 kb (so 500 kb window is drawn arund each TSS).}

\item{downstream}{Numeric specifying the window size (in base pairs) to pad around downstream of each TSS, to fetch gene's enhancers.
Default is 250 kb (so 500 kb window is drawn arund each TSS).}

\item{verbose}{Logical, should warning and info messages be printed?}
}
\value{
A list with enhancers, promoters, and Hi-C matrices for each \code{focus_gene}
}
\description{
The function obtains enhancers and promoters of each gene, and calculates the bulk average Hi-C matrix of the contact between the gene and enhancers.
}
\examples{
data(multiomic_small)
SCEGdata <- process_data(multiomic_small, k_neigh = 5, max_overlap = 0.5)
fpath <- system.file("extdata", package = "SCEGHiC")
gene <- c("TRABD2A", "GNLY", "MFSD6", "CTLA4", "LCLAT1", "NCK2", "GALM", "TMSB10", "ID2", "CXCR4")
weight <- calculateHiCWeights(SCEGdata, species = "Homo sapiens", genome = "hg38", focus_gene = gene, averHicPath = fpath)
weight[["CTLA4"]]$promoters
weight[["CTLA4"]]$enhancers
head(weight[["CTLA4"]]$contact)
}
