% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{connections_Plot}
\alias{connections_Plot}
\title{Plot SCEG-HiC predicted enhancer-gene links}
\usage{
connections_Plot(
  SCEG_HiC_Result,
  species,
  genome,
  focus_gene,
  cutoff = NULL,
  upstream = 250000,
  downstream = 250000,
  gene_anno = NULL
)
}
\arguments{
\item{SCEG_HiC_Result}{A data.frame containing the output from \code{Run_SCEG_HiC()}.}

\item{species}{Character string specifying the species name. Supported values are "Homo sapiens" or "Mus musculus".}

\item{genome}{Character string specifying the genome assembly. Supported values are "hg38", "hg19", "mm10", or "mm9".}

\item{focus_gene}{A character vector of gene symbols to focus on.}

\item{cutoff}{Threshold for selecting gene-peak pairs. Default is \code{NULL}.
If \code{aggregate = TRUE}, we recommend setting \code{cutoff = 0.01}.
If \code{aggregate = FALSE}, we recommend \code{cutoff = 0.001}.}

\item{upstream}{Numeric specifying the number of base pairs upstream of each TSS to define enhancers. Default is 250,000 bp (250 kb).}

\item{downstream}{Numeric specifying the number of base pairs downstream of each TSS to define enhancers. Default is 250,000 bp (250 kb).}

\item{gene_anno}{A \code{data.frame} or \code{NULL}. If provided, should be compatible with the \code{\link[Gviz]{GeneRegionTrack-class}}.
Column names must not contain \code{NA}.}
}
\value{
A genome browser-style plot displaying the focused gene and predicted enhancer-gene connections.
}
\description{
Visualizes the predicted enhancer-gene links from SCEG-HiC for a given gene by plotting them in a genome region view.
}
\examples{
#' data(multiomic_small)
SCEGdata <- process_data(multiomic_small, k_neigh = 5, max_overlap = 0.5)
fpath <- system.file("extdata", package = "SCEGHiC")
gene <- c("TRABD2A", "GNLY", "MFSD6", "CTLA4", "LCLAT1", "NCK2", "GALM", "TMSB10", "ID2", "CXCR4")
weight <- calculateHiCWeights(SCEGdata, species = "Homo sapiens", genome = "hg38", focus_gene = gene, averHicPath = fpath)
results_SCEGHiC <- Run_SCEG_HiC(SCEGdata, weight, focus_gene = gene)
connections_Plot(results_SCEGHiC, species = "Homo sapiens", genome = "hg38", focus_gene = "CTLA4", cutoff = 0.01, gene_anno = NULL)
}
