% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{coverPlot}
\alias{coverPlot}
\title{Plot links of SCEG-HiC on focused genes and Tn5 insertion frequency over a region}
\usage{
coverPlot(
  object,
  focus_gene,
  species,
  genome,
  assay = NULL,
  upstream = 250000,
  downstream = 250000,
  HIC_Result = NULL,
  HIC_cutoff = 0,
  SCEG_HiC_Result = NULL,
  SCEG_HiC_cutoff = 0,
  correlation = NULL,
  cells = NULL,
  cellnames = NULL
)
}
\arguments{
\item{object}{A Seurat object.}

\item{focus_gene}{The focused a gene.}

\item{species}{String indicating the name of the species.It can be either "Homo sapiens","Mus musculus".}

\item{genome}{String indicating the name of the genome assembly.It can be either "hg38",
"hg19", "mm10", "mm9".}

\item{assay}{Name of the assay to plot. If a list of assays is provided,  data from each assay will be
shown overlaid on each track. The first assay in the list will define the assay used for gene annotations,
links, and peaks (if shown). The order of assays given defines the plotting order.}

\item{upstream}{Numeric specifying the window size (in base pairs) to pad around upstream of each TSS, to fetch gene's enhancers.
Default is 250 kb (so 500 kb window is drawn arund each TSS).}

\item{downstream}{Numeric specifying the window size (in base pairs) to pad around downstream of each TSS, to fetch gene's enhancers.
Default is 250 kb (so 500 kb window is drawn arund each TSS).}

\item{HIC_Result}{Hi-C result.}

\item{HIC_cutoff}{The threshold of gene-peak validated by Hi-C.Default is 0.}

\item{SCEG_HiC_Result}{SCEG_HiC result.}

\item{SCEG_HiC_cutoff}{The threshold of gene-peak predicted by SCEG-HiC.Default is 0.If \code{"aggregate"} is TRUE ,we recommend the cutoff of 0.01.
If \code{"aggregate"} is FALSE ,we recommend the cutoff of 0.001.}

\item{correlation}{Correlation between elements result.}

\item{cells}{Which cells to plot. Default all cells.}

\item{cellnames}{Name of one or more metadata columns to group (color) the cells by. Default is the current cell identities.}
}
\value{
Returns a \code{\link[patchwork]{patchwork}} object
}
\description{
Plot frequency of Tn5 insertion events for different groups of cells within
given regions of the genome.  And plot links for gene-peak  predicted by
SCEG-HiC within given regions of the genome.
}
\details{
Additional information can be layered on the coverage plot by setting several
different options in the CoveragePlot function. This includes showing:
\itemize{
\item{gene annotations}
\item{peak positions}
\item{gene-peak results predicted by correlation}
\item{gene-peak results validated by Hi-C}
}
}
\examples{
data(multiomic_small)
SCEGdata <- process_data(multiomic_small, k_neigh = 5, max_overlap = 0.5)
fpath <- system.file("extdata", package = "SCEGHiC")
gene <- c("TRABD2A", "GNLY", "MFSD6", "CTLA4", "LCLAT1", "NCK2", "GALM", "TMSB10", "ID2", "CXCR4")
weight <- calculateHiCWeights(SCEGdata, species = "Homo sapiens", genome = "hg38", focus_gene = gene, averHicPath = fpath)
results_SCEGHiC <- Run_SCEG_HiC(SCEGdata, weight, focus_gene = gene)
fpath <- system.file("extdata", "multiomic_small_atac_fragments.tsv.gz", package = "SCEGHiC")
library(Signac)
frags <- CreateFragmentObject(path = fpath, cells = colnames(multiomic_small))
Fragments(multiomic_small) <- frags
coverPlot(multiomic_small, focus_gene = "CTLA4", species = "Homo sapiens", genome = "hg38", assay = "peaks", SCEG_HiC_Result = results_SCEGHiC, SCEG_HiC_cutoff = 0.01)
}
