% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{liftOverPeaks}
\alias{liftOverPeaks}
\title{Convert genomic coordinates of peaks between genome assemblies}
\usage{
liftOverPeaks(split_peaks, genome)
}
\arguments{
\item{split_peaks}{A matrix or data.frame with separated peak coordinates, typically obtained by splitting peak strings (e.g., "chr2-973888-974755") into columns for chromosome, start, and end.}

\item{genome}{Character string specifying the genome assembly. Supported values are "hg19" or "mm9".}
}
\value{
A data.frame containing peaks after genome coordinate conversion, with updated chromosome, start, and end positions.
}
\description{
This function converts genomic coordinates for a set of peaks to a specified genome assembly.
The \code{SCEGHiC} package provides bulk average Hi-C data for the hg38 and mm10 assemblies,
so coordinate conversion is required for other assemblies such as hg19 and mm9.
}
\examples{
peaks <- c(
  "chr2-973888-974755", "chr2-3195868-3196264", "chr2-3242398-3243508", "chr2-6863856-6866834", "chr2-6913141-6914394",
  "chr2-6924663-6924972", "chr2-6999095-6999554", "chr2-7007713-7008819", "chr2-7083945-7084032", "chr2-7097148-7098029"
)
split_peaks <- do.call("rbind", strsplit(as.character(peaks), "-", fixed = TRUE))
peakinfo <- liftOverPeaks(split_peaks, "hg19")
head(peakinfo)

}
