% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{LinksPlot}
\alias{LinksPlot}
\title{Plot linked genomic elements}
\usage{
LinksPlot(links, geneanno, gene, region, lowcolor, highcolor, titlename)
}
\arguments{
\item{links}{A data.frame containing enhancer-gene links.}

\item{gene}{A character string specifying the target gene to highlight.}

\item{region}{A \code{\link[GenomicRanges]{GRanges}} object specifying the genomic region to plot.}

\item{lowcolor}{Color for the lowest scores in the gradient.}

\item{highcolor}{Color for the highest scores in the gradient.}

\item{titlename}{A character string specifying the title of the plot.}

\item{gene_anno}{A \code{data.frame} or \code{NULL}. If provided, should be compatible with the \code{\link[Gviz]{GeneRegionTrack-class}}.
Column names must not contain \code{NA}.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object visualizing enhancer-gene links.
}
\description{
Visualize predicted links between genomic elements (e.g., genes and regulatory peaks) within a specified genomic region.
}
\examples{
links <- data.frame(
  gene = "CTLA4", peak = c("chr2_203623664_203623982", "chr2_203730093_203731243", "chr2_203992800_203993979"),
  score = c(0.03623216, 0.14814205, 0.43240254)
)
library(GenomicRanges)
region <- GRanges(seqnames = Rle("chr2"), ranges = IRanges(start = 203617771, end = 204117772))
geneanno <- annotateTSS("Homo sapiens", "hg38")
colnames(geneanno) <- c("chr", "gene", "tss")
links <- LinksPlot(links, geneanno, gene = "CTLA4", region, lowcolor = "blue", highcolor = "orange", titlename = "SCEG-HiC")
links
}
