% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{LinksPlot}
\alias{LinksPlot}
\title{Plot linked genomic elements}
\usage{
LinksPlot(links, geneanno, gene, region, lowcolor, highcolor, titlename)
}
\arguments{
\item{links}{The result of gene-peak pairs.}

\item{geneanno}{Either \code{NULL} or a data.frame. The data.frame should be in a form compatible with the Gviz function.
\code{\link[Gviz]{GeneRegionTrack-class}} (cannot have NA as column names).}

\item{gene}{The focused genes.}

\item{region}{A genomic region to plot.}

\item{lowcolor}{Colours for low ends of the gradient.}

\item{highcolor}{Colours for high ends of the gradient.}

\item{titlename}{The title of linking genomic elements.}
}
\value{
Returns a \code{\link[ggplot2]{ggplot}} object
}
\description{
Display links between pairs of genomic elements within a given region of the genome.
}
\examples{
links <- data.frame(
  gene = "CTLA4", peak = c("chr2_203623664_203623982", "chr2_203730093_203731243", "chr2_203992800_203993979"),
  score = c(0.03623216, 0.14814205, 0.43240254)
)
library(GenomicRanges)
region <- GRanges(seqnames = Rle("chr2"), ranges = IRanges(start = 203617771, end = 204117772))
geneanno <- annotateTSS("Homo sapiens", "hg38")
colnames(geneanno) <- c("chr", "gene", "tss")
links <- LinksPlot(links, geneanno, gene = "CTLA4", region, lowcolor = "blue", highcolor = "orange", titlename = "SCEG-HiC")
links
}
