% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{normalizeHiCWeights}
\alias{normalizeHiCWeights}
\title{The bulk average Hi-C matrix normalization}
\usage{
normalizeHiCWeights(hicweight, normalizeMethod = "1", element)
}
\arguments{
\item{hicweight}{Raw the bulk average Hi-C matrix.}

\item{normalizeMethod}{The method of normalizing the average Hi-C.Default is "1".
\itemize{
\item \code{"1"}, rank score to normalize the average Hi-C.
\item \code{"2"}, min-max to normalize the average Hi-C.
\item \code{"3"}, 0-1 to normalize the average Hi-C.
\item \code{"4"}, -log10 to normalize the average Hi-C.
}}

\item{element}{The row or col elements of the average Hi-C matrix.}
}
\value{
A matrix with normalizing average Hi-C.
}
\description{
The bulk average Hi-C matrix normalization
}
\examples{
hicweight <- data.frame(
  element1 = c(
    "CHN1", "CHN1", "CHN1", "CHN1", "CHN1", "chr2_174901074_174901475", "chr2_174901074_174901475",
    "chr2_174901074_174901475", "chr2_174901074_174901475", "chr2_174990102_174990619", "chr2_174990102_174990619",
    "chr2_174990102_174990619", "chr2_175048293_175048704", "chr2_175048293_175048704", "chr2_175167038_175169195"
  ),
  element2 = c(
    "chr2_174901074_174901475", "chr2_174990102_174990619", "chr2_175048293_175048704", "chr2_175167038_175169195",
    "chr2_175187467_175187717", "chr2_174990102_174990619", "chr2_175048293_175048704", "chr2_175167038_175169195",
    "chr2_175187467_175187717", "chr2_175048293_175048704", "chr2_175167038_175169195", "chr2_175187467_175187717",
    "chr2_175167038_175169195", "chr2_175187467_175187717", "chr2_175187467_175187717"
  ),
  score = c(
    0.0016062358, 0.0076253167, 0.0031021570, 0.0012148262, 0.0010788392, 0.0019968580, 0.0012957845,
    0.0009198567, 0.0011448870, 0.0023991885, 0.0010785481, 0.0009664508, 0.0012865316, 0.0012122647, 0.0057781939
  )
)
element <- unique(c(hicweight$element1, hicweight$element2))
weight <- normalizeHiCWeights(hicweight, "1", element)
head(weight)
}
