% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{Run_SCEG_HiC}
\alias{Run_SCEG_HiC}
\title{Run SCEG-HiC's model}
\usage{
Run_SCEG_HiC(
  SCEGdata,
  HiCWeights,
  method = "wglasso",
  focus_gene,
  normalizeMethod = "1",
  cutoff = NULL,
  verbose = TRUE
)
}
\arguments{
\item{SCEGdata}{Preprocessed data for SCEG-HiC.}

\item{HiCWeights}{Raw the bulk average Hi-C matrix.Run the result of code{'calculateHiCWeights'}}

\item{method}{The method of gene network reconstruction.Default is wglasso.}

\item{focus_gene}{The focused genes.}

\item{normalizeMethod}{The method of normalizing the average Hi-C.Default is "1".
\itemize{
\item \code{'1'},-rank score to normalize the average Hi-C.
\item \code{'2'},-min-max to normalize the average Hi-C.
\item \code{'3'},-0-1 to normalize the average Hi-C.
\item \code{'4'},--log10 to normalize the average Hi-C.
}}

\item{cutoff}{The threshold of gene-peak.Default is NULL.If \code{"aggregate"} is TRUE ,we recommend the cutoff of 0.01.
If \code{"aggregate"} is FALSE ,we recommend the cutoff of 0.001.}

\item{verbose}{Logical, should warning and info messages be printed?}
}
\value{
A data.frame of correlations for each gene-peak pair.
}
\description{
The model uses data preprocessed by SCEG-HIC, the bulk average Hi-C as the penalty matrix,
calculates the partial correlation coefficient matrix between elements using the wglasso model,
and selects the gene-peak.
}
\examples{
data(multiomic_small)
SCEGdata <- process_data(multiomic_small, k_neigh = 5, max_overlap = 0.5)
fpath <- system.file("extdata", package = "SCEGHiC")
gene <- c("TRABD2A", "GNLY", "MFSD6", "CTLA4", "LCLAT1", "NCK2", "GALM", "TMSB10", "ID2", "CXCR4")
weight <- calculateHiCWeights(SCEGdata, species = "Homo sapiens", genome = "hg38", focus_gene = gene, averHicPath = fpath)
results_SCEGHiC <- Run_SCEG_HiC(SCEGdata, weight, focus_gene = gene)
results_SCEGHiC[results_SCEGHiC$gene == "CTLA4", ]

}
