% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{wglasso}
\alias{wglasso}
\title{Gene network reconstruction by integration of prior biological knowledge}
\usage{
wglasso(S, sample, normhicMatrix, element)
}
\arguments{
\item{S}{Covariance matrix:p by p matrix (symmetric)}

\item{sample}{The number of samples.}

\item{normhicMatrix}{A matrix with normalizing average Hi-C.}

\item{element}{The row or col elements of the average Hi-C matrix.}
}
\value{
A matrix of partial correlation coefficients between elements.
}
\description{
The method, instead of using a single penalty, different penalty values were applied for element pairs
based on a priori knowledge as to whether the two element should be connected. The a priori information
can be calculated or retrieved from other biological data, e.g., the bulk average Hi-C matrix.
}
